Refactor core to base, update scripts and dependencies
