+++ packages/components/base/Collapsible/description/index.ts
