+++ packages/components/base/DialogBox/variants/index.ts
