+++ packages/components/base/Pagination/variants/index.ts
