+++ packages/components/base/Switch/variants/index.ts
