+++ packages/components/base/Table/src.tsx
@@ -1,6 +1,6 @@
 import React from "react";
 import { cn } from "@/lib/utils";
-import { CheckBox } from "@repo/components/core/Checkbox/src";
+import { CheckBox } from "@repo/components/base/Checkbox/src";
 
 interface Column {
   head: string;
