+++ packages/components/base/Table/variants/index.ts
