+++ packages/components/index.tsx
@@ -1,3 +1,3 @@
 // Barrel file
 
-export * from "./core/__template";
+export * from "./base/__template";
