+++ packages/components/tsconfig.json
@@ -8,7 +8,7 @@
     ],
     "baseUrl": ".",
     "paths": {
-      "@/core/*": ["core/*"],
+      "@/core/*": ["base/*"],
       "@/lib/*": ["lib/*"]
     }
   },
